rename: Reanme SkeletonUI => TableSkeletonUI
